Added tests for UIButtonRenderer
The tests are split into 4 test classes.

`NUIButtonRendererTests` - the test class for the majority of
properties.
`NUIButtonRendererBackgroundColorTests` - a class which tests the
setting of background colors.
`NUIButtonRendererBackgroundImageTests` - a class which tests the
setting of background images.
`NUIButtonRendererForegroundImageTests` - a class which tests the
setting of foreground images.
`NUIButtonGradientSnapshotTest` - a subclass of `FBSnapshotTestCase`
which checks that gradients are applied correctly.