Install cocoapods via Bundler to lock version and resolve travis issues
When running gem install cocoapods, for some reason a beta version of
activesupport (a dependency of cocoapods) was attempted to be installed.
This beta has higher ruby version requirement, that is not satisfied by some of
the Travis images' defaults.

When installing Travis via bundler, the stable activesupport version is chosen
and now locked via the Gemfile.lock